--- src/main/java/com/seniors/api/post/repository/PostRepository.java
@@ -1,8 +1,0 @@
-package com.seniors.api.post.repository;
-
-
-import com.seniors.api.post.domain.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<Post, Long>, PostRepositoryCustom {
-}
