--- src/main/java/com/seniors/api/post/repository/PostRepositoryCustom.java
@@ -1,8 +1,0 @@
-package com.seniors.api.post.repository;
-
-import com.seniors.api.post.domain.Post;
-
-public interface PostRepositoryCustom {
-
-	Post getOnePost(Long postId);
-}
