--- src/main/java/com/seniors/common/exception/ConflictException.java
@@ -1,19 +1,0 @@
-package com.seniors.common.exception;
-
-public class ConflictException extends RuntimeException {
-
-	public ConflictException() {
-	}
-
-	public ConflictException(Throwable throwable) {
-		super(throwable);
-	}
-
-	public ConflictException(String message) {
-		super(message);
-	}
-
-	public ConflictException(String message, Throwable throwable) {
-		super(message, throwable);
-	}
-}
