--- src/main/java/com/seniors/common/exception/FailEncryptException.java
@@ -1,15 +1,0 @@
-package com.seniors.common.exception;
-
-public class FailEncryptException extends RuntimeException {
-
-	public FailEncryptException() {
-	}
-
-	public FailEncryptException(Throwable throwable) {
-		super(throwable);
-	}
-
-	public FailEncryptException(String message) {
-		super(message);
-	}
-}
