--- src/main/java/com/seniors/common/exception/FileUploadException.java
@@ -1,19 +1,0 @@
-package com.seniors.common.exception;
-
-public class FileUploadException extends RuntimeException {
-
-	public FileUploadException() {
-	}
-
-	public FileUploadException(Throwable throwable) {
-		super(throwable);
-	}
-
-	public FileUploadException(String message) {
-		super(message);
-	}
-
-	public FileUploadException(String message, Throwable throwable) {
-		super(message, throwable);
-	}
-}
