--- src/main/java/com/seniors/common/exception/NotAuthorizedException.java
@@ -1,19 +1,0 @@
-package com.seniors.common.exception;
-
-public class NotAuthorizedException extends RuntimeException {
-
-	public NotAuthorizedException() {
-	}
-
-	public NotAuthorizedException(Throwable throwable) {
-		super(throwable);
-	}
-
-	public NotAuthorizedException(String message) {
-		super(message);
-	}
-
-	public NotAuthorizedException(String message, Throwable throwable) {
-		super(message, throwable);
-	}
-}
