--- src/main/java/com/seniors/common/exception/ViewException.java
@@ -1,19 +1,0 @@
-package com.seniors.common.exception;
-
-public class ViewException extends RuntimeException{
-
-	public ViewException() {
-	}
-
-	public ViewException(Throwable throwable) {
-		super(throwable);
-	}
-
-	public ViewException(String message) {
-		super(message);
-	}
-
-	public ViewException(String message, Throwable throwable) {
-		super(message, throwable);
-	}
-}
