+++ src/main/java/com/seniors/common/exception/type/ConflictException.java
@@ -1,0 +1,19 @@
+package com.seniors.common.exception.type;
+
+public class ConflictException extends RuntimeException {
+
+	public ConflictException() {
+	}
+
+	public ConflictException(Throwable throwable) {
+		super(throwable);
+	}
+
+	public ConflictException(String message) {
+		super(message);
+	}
+
+	public ConflictException(String message, Throwable throwable) {
+		super(message, throwable);
+	}
+}
