+++ src/main/java/com/seniors/common/exception/type/FailEncryptException.java
@@ -1,0 +1,15 @@
+package com.seniors.common.exception.type;
+
+public class FailEncryptException extends RuntimeException {
+
+	public FailEncryptException() {
+	}
+
+	public FailEncryptException(Throwable throwable) {
+		super(throwable);
+	}
+
+	public FailEncryptException(String message) {
+		super(message);
+	}
+}
