+++ src/main/java/com/seniors/common/exception/type/NotAuthorizedException.java
@@ -1,0 +1,19 @@
+package com.seniors.common.exception.type;
+
+public class NotAuthorizedException extends RuntimeException {
+
+	public NotAuthorizedException() {
+	}
+
+	public NotAuthorizedException(Throwable throwable) {
+		super(throwable);
+	}
+
+	public NotAuthorizedException(String message) {
+		super(message);
+	}
+
+	public NotAuthorizedException(String message, Throwable throwable) {
+		super(message, throwable);
+	}
+}
