+++ src/main/java/com/seniors/common/exception/type/ViewException.java
@@ -1,0 +1,19 @@
+package com.seniors.common.exception.type;
+
+public class ViewException extends RuntimeException{
+
+	public ViewException() {
+	}
+
+	public ViewException(Throwable throwable) {
+		super(throwable);
+	}
+
+	public ViewException(String message) {
+		super(message);
+	}
+
+	public ViewException(String message, Throwable throwable) {
+		super(message, throwable);
+	}
+}
