+++ src/main/java/com/seniors/common/exception/type/post/PostNotFound.java
@@ -1,0 +1,20 @@
+package com.seniors.common.exception.type.post;
+
+import com.seniors.common.exception.type.CustomException;
+
+public class PostNotFound extends CustomException {
+	public PostNotFound() {
+	}
+
+	public PostNotFound(Throwable throwable) {
+		super(throwable);
+	}
+
+	public PostNotFound(String message) {
+		super(message);
+	}
+
+	public PostNotFound(String message, Throwable throwable) {
+		super(message, throwable);
+	}
+}
