+++ src/main/java/com/seniors/domain/common/BaseEntity.java
@@ -1,0 +1,24 @@
+package com.seniors.domain.common;
+
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+
+@EntityListeners(AuditingEntityListener.class)
+@MappedSuperclass
+@Getter
+public class BaseEntity extends BaseTimeEntity{
+
+	@CreatedBy
+	@Column(updatable = false)
+	private String createdBy;
+
+	@Column(insertable = false)
+	@LastModifiedBy
+	private String lastModifiedBy;
+}
