+++ src/main/java/com/seniors/domain/post/repository/PostRepository.java
@@ -1,0 +1,8 @@
+package com.seniors.domain.post.repository;
+
+
+import com.seniors.domain.post.entity.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post, Long>, PostRepositoryCustom {
+}
