+++ src/main/java/com/seniors/domain/post/repository/PostRepositoryCustom.java
@@ -1,0 +1,8 @@
+package com.seniors.domain.post.repository;
+
+import com.seniors.domain.post.entity.Post;
+
+public interface PostRepositoryCustom {
+
+	Post getOnePost(Long postId);
+}
