--- src/test/java/com/seniors/api/comment/CommentController.java
@@ -1,4 +1,0 @@
-package com.seniors.api.comment;
-
-public class CommentController {
-}
