--- src/test/java/com/seniors/api/post/PostControllerTest.java
@@ -1,4 +1,0 @@
-package com.seniors.api.post;
-
-public class PostControllerTest {
-}
