+++ src/test/java/com/seniors/domain/comment/CommentController.java
@@ -1,0 +1,4 @@
+package com.seniors.domain.comment;
+
+public class CommentController {
+}
