+++ src/test/java/com/seniors/domain/post/PostControllerTest.java
@@ -1,0 +1,4 @@
+package com.seniors.domain.post;
+
+public class PostControllerTest {
+}
